python_modules/requirements.txt
requests
beautifulsoup4
openpyxl
mysql-connector-python
selenium
webdriver-manager
aiogram==3.4.1
python-dotenv
psutil
